[CL 2594426 by Peter Knepley in UT-Release branch]

+++ Engine/Source/Programs/AutomationTool/Mac/MacPlatform.Automation.cs
@@ -16,9 +16,22 @@ public class MacPlatform : Platform
 
 	public override string GetCookPlatform(bool bDedicatedServer, bool bIsClientOnly, string CookFlavor)
 	{
-		const string ClientCookPlatform = "MacNoEditor";
+		const string NoEditorCookPlatform = "MacNoEditor";
 		const string ServerCookPlatform = "MacServer";
-		return bDedicatedServer ? ServerCookPlatform : ClientCookPlatform;
+		const string ClientCookPlatform = "MacClient";
+
+		if (bDedicatedServer)
+		{
+			return ServerCookPlatform;
+		}
+		else if (bIsClientOnly)
+		{
+			return ClientCookPlatform;
+		}
+		else
+		{
+			return NoEditorCookPlatform;
+		}
 	}
 
 	public override string GetEditorCookPlatform()
@@ -26,13 +39,14 @@ public class MacPlatform : Platform
 		return "Mac";
 	}
 
-	private void StageAppBundle(DeploymentContext SC, string InPath, string NewName)
+	private void StageAppBundle(DeploymentContext SC, StagedFileType InStageFileType, string InPath, string NewName)
 	{
-		SC.StageFiles(StagedFileType.NonUFS, InPath, "*", true, null, NewName, false, true, null);
+		SC.StageFiles(InStageFileType, InPath, "*", true, null, NewName, false, true, null);
 	}
 
 	public override void GetFilesToDeployOrStage(ProjectParams Params, DeploymentContext SC)
 	{
+	
 		if (Params.StageNonMonolithic)
 		{
 			if (SC.DedicatedServer)
@@ -61,17 +75,17 @@ public class MacPlatform : Platform
 				if (SC.StageTargetConfigurations.Contains(UnrealTargetConfiguration.Development))
 				{
 					SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.LocalRoot, "Engine/Binaries", SC.PlatformDir, "UE4.app"));
-					SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.ProjectRoot, "Binaries", SC.PlatformDir), "UE4-" + SC.ShortProjectName + ".dylib");
+					SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.ProjectRoot, "Binaries", SC.PlatformDir), "*.dylib");
 				}
 				if (SC.StageTargetConfigurations.Contains(UnrealTargetConfiguration.Test))
 				{
 					SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.LocalRoot, "Engine/Binaries", SC.PlatformDir, "UE4-Mac-Test.app"));
-					SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.ProjectRoot, "Binaries", SC.PlatformDir), "UE4-" + SC.ShortProjectName + "-Mac-Test.dylib");
+					SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.ProjectRoot, "Binaries", SC.PlatformDir), "*-Mac-Test.dylib");
 				}
 				if (SC.StageTargetConfigurations.Contains(UnrealTargetConfiguration.Shipping))
 				{
 					SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.LocalRoot, "Engine/Binaries", SC.PlatformDir, "UE4-Mac-Shipping.app"));
-					SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.ProjectRoot, "Binaries", SC.PlatformDir), "UE4-" + SC.ShortProjectName + "-Mac-Shipping.dylib");
+					SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.ProjectRoot, "Binaries", SC.PlatformDir), "*-Mac-Shipping.dylib");
 				}
 
 				SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.LocalRoot, "Engine/Plugins"), "UE4-*.dylib", true);
@@ -80,6 +94,9 @@ public class MacPlatform : Platform
 		}
 		else
 		{
+			// the first app is the "main" one, the rest are marked as debug files for exclusion from chunking/distribution
+			StagedFileType WorkingFileType = StagedFileType.NonUFS;
+
 			List<string> Exes = GetExecutableNames(SC);
 			foreach (var Exe in Exes)
 			{
@@ -87,7 +104,7 @@ public class MacPlatform : Platform
 				if (Exe.StartsWith(CombinePaths(SC.RuntimeProjectRootDir, "Binaries", SC.PlatformDir)))
 				{
 					AppBundlePath = CombinePaths(SC.ShortProjectName, "Binaries", SC.PlatformDir, Path.GetFileNameWithoutExtension(Exe) + ".app");
-					StageAppBundle(SC, CombinePaths(SC.ProjectRoot, "Binaries", SC.PlatformDir, Path.GetFileNameWithoutExtension(Exe) + ".app"), AppBundlePath);
+					StageAppBundle(SC, WorkingFileType, CombinePaths(SC.ProjectRoot, "Binaries", SC.PlatformDir, Path.GetFileNameWithoutExtension(Exe) + ".app"), AppBundlePath);
 				}
 				else if (Exe.StartsWith(CombinePaths(SC.RuntimeRootDir, "Engine/Binaries", SC.PlatformDir)))
 				{
@@ -95,30 +112,40 @@ public class MacPlatform : Platform
 
 					string AbsoluteBundlePath = CombinePaths (SC.LocalRoot, AppBundlePath);
 					// ensure the ue4game binary exists, if applicable
-					if (!SC.IsCodeBasedProject && !Directory.Exists(AbsoluteBundlePath))
+					if (!SC.IsCodeBasedProject && !Directory.Exists(AbsoluteBundlePath) && !SC.bIsCombiningMultiplePlatforms)
 					{
 						Log("Failed to find app bundle " + AbsoluteBundlePath);
 						AutomationTool.ErrorReporter.Error("Stage Failed.", (int)AutomationTool.ErrorCodes.Error_MissingExecutable);
 						throw new AutomationException("Could not find app bundle {0}. You may need to build the UE4 project with your target configuration and platform.", AbsoluteBundlePath);
 					}
 
-					StageAppBundle(SC, CombinePaths(SC.LocalRoot, "Engine/Binaries", SC.PlatformDir, Path.GetFileNameWithoutExtension(Exe) + ".app"), AppBundlePath);
+					StageAppBundle(SC, WorkingFileType, CombinePaths(SC.LocalRoot, "Engine/Binaries", SC.PlatformDir, Path.GetFileNameWithoutExtension(Exe) + ".app"), AppBundlePath);
 				}
 
 				if (!string.IsNullOrEmpty(AppBundlePath))
 				{
-					SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.ProjectRoot, "Build/Mac"), "Application.icns", false, null, CombinePaths(AppBundlePath, "Contents/Resources"), true);
+					SC.StageFiles(WorkingFileType, CombinePaths(SC.ProjectRoot, "Build/Mac"), "Application.icns", false, null, CombinePaths(AppBundlePath, "Contents/Resources"), true);
+
+					if (Params.bUsesSteam)
+					{
+						SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.LocalRoot, "Engine/Source/ThirdParty/Steamworks/Steamv132/sdk/redistributable_bin/osx32"), "libsteam_api.dylib", false, null, CombinePaths(AppBundlePath, "Contents/MacOS"), true);
+					}
 				}
+
+				// the first app is the "main" one, the rest are marked as debug files for exclusion from chunking/distribution
+				WorkingFileType = StagedFileType.DebugNonUFS;
 			}
 		}
-
 		// Copy the splash screen, Mac specific
 		SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.ProjectRoot, "Content/Splash"), "Splash.bmp", false, null, null, true);
 
 		// CEF3 files
-		SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.LocalRoot, "Engine/Binaries/ThirdParty/CEF3/Mac/"), "*", true, null, null, true);
-		string UnrealCEFSubProcessPath = CombinePaths("Engine/Binaries", SC.PlatformDir, "UnrealCEFSubProcess.app");
-		StageAppBundle(SC, CombinePaths(SC.LocalRoot, "Engine/Binaries", SC.PlatformDir, "UnrealCEFSubProcess.app"), UnrealCEFSubProcessPath);
+		if(Params.bUsesCEF3)
+		{
+			SC.StageFiles(StagedFileType.NonUFS, CombinePaths(SC.LocalRoot, "Engine/Binaries/ThirdParty/CEF3/Mac/"), "*", true, null, null, true);
+			string UnrealCEFSubProcessPath = CombinePaths("Engine/Binaries", SC.PlatformDir, "UnrealCEFSubProcess.app");
+			StageAppBundle(SC, StagedFileType.NonUFS, CombinePaths(SC.LocalRoot, "Engine/Binaries", SC.PlatformDir, "UnrealCEFSubProcess.app"), UnrealCEFSubProcessPath);
+		}
 	}
 
 	public override void Package(ProjectParams Params, DeploymentContext SC, int WorkingCL)
@@ -131,54 +158,69 @@ public class MacPlatform : Platform
 	{
 		string ExeName = SC.StageExecutables[0];
 		string BundlePath = CombinePaths(SC.ArchiveDirectory, SC.ShortProjectName + ".app");
-		if (!Directory.Exists(BundlePath))
+
+		if (SC.bIsCombiningMultiplePlatforms)
 		{
-			string SourceBundlePath = CombinePaths(SC.ArchiveDirectory, SC.ShortProjectName, "Binaries", "Mac", ExeName + ".app");
-			if (!Directory.Exists(SourceBundlePath))
+			// when combining multiple platforms, don't merge the content into the .app, use the one in the Binaries directory
+			BundlePath = CombinePaths(SC.ArchiveDirectory, SC.ShortProjectName, "Binaries", "Mac", ExeName + ".app");
+			if (!Directory.Exists(BundlePath))
 			{
-				SourceBundlePath = CombinePaths(SC.ArchiveDirectory, "Engine", "Binaries", "Mac", ExeName + ".app");
+				// if the .app wasn't there, just skip out (we don't require executables when combining)
+				return;
 			}
-			Directory.Move(SourceBundlePath, BundlePath);
 		}
 
 		string TargetPath = CombinePaths(BundlePath, "Contents", "UE4");
-		if (DirectoryExists(TargetPath))
+		if (!SC.bIsCombiningMultiplePlatforms)
 		{
-			Directory.Delete(TargetPath, true);
-		}
-
-		// First, move all files and folders inside he app bundle
-		string[] StagedFiles = Directory.GetFiles(SC.ArchiveDirectory, "*", SearchOption.TopDirectoryOnly);
-		foreach (string FilePath in StagedFiles)
-		{
-			string TargetFilePath = CombinePaths(TargetPath, Path.GetFileName(FilePath));
-			Directory.CreateDirectory(Path.GetDirectoryName(TargetFilePath));
-			File.Move(FilePath, TargetFilePath);
-		}
-
-		string[] StagedDirectories = Directory.GetDirectories(SC.ArchiveDirectory, "*", SearchOption.TopDirectoryOnly);
-		foreach (string DirPath in StagedDirectories)
-		{
-			string DirName = Path.GetFileName(DirPath);
-			if (!DirName.EndsWith(".app"))
+			if (!Directory.Exists(BundlePath))
 			{
-				string TargetDirPath = CombinePaths(TargetPath, DirName);
-				Directory.CreateDirectory(Path.GetDirectoryName(TargetDirPath));
-				Directory.Move(DirPath, TargetDirPath);
+				string SourceBundlePath = CombinePaths(SC.ArchiveDirectory, SC.ShortProjectName, "Binaries", "Mac", ExeName + ".app");
+				if (!Directory.Exists(SourceBundlePath))
+				{
+					SourceBundlePath = CombinePaths(SC.ArchiveDirectory, "Engine", "Binaries", "Mac", ExeName + ".app");
+				}
+				Directory.Move(SourceBundlePath, BundlePath);
+			}
+
+			if (DirectoryExists(TargetPath))
+			{
+				Directory.Delete(TargetPath, true);
+			}
+
+			// First, move all files and folders inside he app bundle
+			string[] StagedFiles = Directory.GetFiles(SC.ArchiveDirectory, "*", SearchOption.TopDirectoryOnly);
+			foreach (string FilePath in StagedFiles)
+			{
+				string TargetFilePath = CombinePaths(TargetPath, Path.GetFileName(FilePath));
+				Directory.CreateDirectory(Path.GetDirectoryName(TargetFilePath));
+				File.Move(FilePath, TargetFilePath);
+			}
+
+			string[] StagedDirectories = Directory.GetDirectories(SC.ArchiveDirectory, "*", SearchOption.TopDirectoryOnly);
+			foreach (string DirPath in StagedDirectories)
+			{
+				string DirName = Path.GetFileName(DirPath);
+				if (!DirName.EndsWith(".app"))
+				{
+					string TargetDirPath = CombinePaths(TargetPath, DirName);
+					Directory.CreateDirectory(Path.GetDirectoryName(TargetDirPath));
+					Directory.Move(DirPath, TargetDirPath);
+				}
 			}
 		}
 
 		// Update executable name, icon and entry in Info.plist
-		string UE4GamePath = CombinePaths(SC.ArchiveDirectory, SC.ShortProjectName + ".app", "Contents", "MacOS", ExeName);
+		string UE4GamePath = CombinePaths(BundlePath, "Contents", "MacOS", ExeName);
 		if (ExeName != SC.ShortProjectName && File.Exists(UE4GamePath))
 		{
-			string GameExePath = CombinePaths(SC.ArchiveDirectory, SC.ShortProjectName + ".app", "Contents", "MacOS", SC.ShortProjectName);
+			string GameExePath = CombinePaths(BundlePath, "Contents", "MacOS", SC.ShortProjectName);
 			File.Delete(GameExePath);
 			File.Move(UE4GamePath, GameExePath);
 
-			string DefaultIconPath = CombinePaths(SC.ArchiveDirectory, SC.ShortProjectName + ".app", "Contents", "Resources", "UE4.icns");
-			string CustomIconSrcPath = CombinePaths(SC.ArchiveDirectory, SC.ShortProjectName + ".app", "Contents", "Resources", "Application.icns");
-			string CustomIconDestPath = CombinePaths(SC.ArchiveDirectory, SC.ShortProjectName + ".app", "Contents", "Resources", SC.ShortProjectName + ".icns");
+			string DefaultIconPath = CombinePaths(BundlePath, "Contents", "Resources", "UE4.icns");
+			string CustomIconSrcPath = CombinePaths(BundlePath, "Contents", "Resources", "Application.icns");
+			string CustomIconDestPath = CombinePaths(BundlePath, "Contents", "Resources", SC.ShortProjectName + ".icns");
 			if (File.Exists(CustomIconSrcPath))
 			{
 				File.Delete(DefaultIconPath);
@@ -189,7 +231,7 @@ public class MacPlatform : Platform
 				File.Move(DefaultIconPath, CustomIconDestPath);
 			}
 
-			string InfoPlistPath = CombinePaths(SC.ArchiveDirectory, SC.ShortProjectName + ".app", "Contents", "Info.plist");
+			string InfoPlistPath = CombinePaths(BundlePath, "Contents", "Info.plist");
 			string InfoPlistContents = File.ReadAllText(InfoPlistPath);
 			InfoPlistContents = InfoPlistContents.Replace(ExeName, SC.ShortProjectName);
 			InfoPlistContents = InfoPlistContents.Replace("<string>UE4</string>", "<string>" + SC.ShortProjectName + "</string>");
@@ -197,7 +239,12 @@ public class MacPlatform : Platform
 			File.WriteAllText(InfoPlistPath, InfoPlistContents);
 		}
 
-		Directory.CreateDirectory(CombinePaths(TargetPath, ExeName.StartsWith("UE4Game") ? "Engine" : SC.ShortProjectName, "Binaries", "Mac"));
+		if (!SC.bIsCombiningMultiplePlatforms)
+		{
+			// creating this directory when the content isn't moved into the application causes it 
+			// to fail to load, and isn't needed
+			Directory.CreateDirectory(CombinePaths(TargetPath, ExeName.StartsWith("UE4Game") ? "Engine" : SC.ShortProjectName, "Binaries", "Mac"));
+		}
 	}
 
 	public override ProcessResult RunClient(ERunOptions ClientRunFlags, string ClientApp, string ClientCmdLine, ProjectParams Params)
